# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta API",
    "description": "示範範例生成文件",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "必須加上 API Token"
    }
  },
  "paths": {
    "/user/create": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "Users - 使用者"
        ],
        "description": "使用者登入頁面 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "Malenia5@eldenring.io"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "為登入使用者產生 token",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "登入成功"
                },
                "data": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYyN2Y2Mjk2ZjkyYWEwZjI5ZTgxM2M2OCIsImlhdCI6MTY1MjUxNTQ3OCwiZXhwIjoxNjUzMTIwMjc4fQ.94SlCEqzEiyEeLS7QcPujjvJj1HvxJgCqQit2m8qgE8"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/update_password": {
      "post": {
        "tags": [
          "Users - 使用者"
        ],
        "description": "使用者變更密碼頁面 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "password1"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "password1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取得使用者ID和姓名",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "成功更新使用者密碼"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "Users - 使用者"
        ],
        "description": "使用者資訊頁面 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "Malenia5"
                }
              },
              "required": [
                "id"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得使用者資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "成功取得使用者資訊"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Malenia"
                    },
                    "gender": {
                      "type": "string",
                      "example": "female"
                    },
                    "photo": {
                      "type": "string",
                      "example": "www.melenia.io/Malenia5"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users - 使用者"
        ],
        "description": "使用者更新頁面 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "資料格式",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Malenia5"
                },
                "photo": {
                  "type": "string",
                  "example": "www@eldenring.io/Malenia5"
                },
                "gender": {
                  "type": "string",
                  "example": "female"
                }
              },
              "required": [
                "name",
                "photo",
                "gender"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功更新使用者資訊",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "成功更新使用者資訊！"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}